Move all worker to a single folder below "workerTaskManager". Update TransportUtils documentation.
