+++ examples/jsm/loaders/workerTaskManager/WorkerTaskManager.js
@@ -4,7 +4,7 @@
  */
 
 import { FileLoader } from "../../../../build/three.module.js";
-import { WorkerTaskManagerDefaultRouting } from "./comm/worker/defaultRouting.js";
+import { WorkerTaskManagerDefaultRouting } from "./worker/defaultRouting.js";
 
 /**
  * Register one to many tasks type to the WorkerTaskManager. Then init and enqueue a worker based execution by passing
