--- examples/jsm/loaders/workerTaskManager/comm/worker/defaultRouting.js
@@ -1,39 +1,0 @@
-/**
- * @author Kai Salmen / www.kaisalmen.de
- */
-
-const WorkerTaskManagerDefaultRouting = {
-
-	comRouting: function ( context, message, object, initFunction, executeFunction ) {
-
-		let payload = message.data;
-		if ( payload.cmd === 'init' ) {
-
-			if ( object !== undefined && object !== null ) {
-
-				object[ initFunction ]( context, payload.workerId, payload.config );
-			} else {
-
-				initFunction( context, payload.workerId, payload.config );
-
-			}
-
-		} else if ( payload.cmd === 'execute' ) {
-
-			if ( object !== undefined && object !== null ) {
-
-				object[ executeFunction ]( context, payload.workerId, payload.config );
-
-			} else {
-
-				executeFunction( context, payload.workerId, payload.config );
-
-			}
-
-		}
-
-	}
-
-};
-
-export { WorkerTaskManagerDefaultRouting }
