+++ examples/jsm/loaders/workerTaskManager/worker/defaultRouting.js
