+++ examples/jsm/loaders/workerTaskManager/worker/tmModuleExampleNoThree.js
@@ -2,8 +2,8 @@
  * @author Kai Salmen / www.kaisalmen.de
  */
 
-import { GeometryTransport } from "../workerTaskManager/utils/TransportUtils.js";
-import { WorkerTaskManagerDefaultRouting } from "../workerTaskManager/comm/worker/defaultRouting.js";
+import { GeometryTransport } from "../utils/TransportUtils.js";
+import { WorkerTaskManagerDefaultRouting } from "./defaultRouting.js";
 
 function init ( context, id, config ) {
